services/financial/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
redis>=4.5.0
pydantic>=2.0.0
httpx>=0.24.0
